Develop role 4

+++ app/Http/Controllers/RoleController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 use App\Models\Role;
 use App\Models\Permission;
@@ -13,7 +14,7 @@ use App\Models\RolePermission;
 class RoleController extends Controller
 {
     public function index()
-    {
+    {   Gate::authorize('viewUsersAndRole', Role::class);
         $pageTitle = 'Role Lists';
         $roles = Role::all();
 
@@ -43,7 +44,7 @@ class RoleController extends Controller
             'permissionIds' => ['required'],
         ]);
 
-        Gate::authorize('createAnyRoles', User::class);
+        Gate::authorize('createNewRoles', Role::class);
 
         DB::beginTransaction();
         try {
@@ -65,6 +66,7 @@ class RoleController extends Controller
     public function edit($id)//
     {
             $pageTitle = 'Edit Role';
+            Gate::authorize('updateAnyRoles', Role::class);
             $permissions = Permission::all();
             $role = Role::findOrFail($id);
             
@@ -75,7 +77,7 @@ class RoleController extends Controller
     public function update(Request $request, $id)//
     {
         $role = Role::findOrFail($id);
-
+        Gate::authorize('updateAnyRoles', Role::class);
 
         DB::beginTransaction();
         try {
@@ -98,6 +100,7 @@ class RoleController extends Controller
     {
         $pageTitle = 'Delete Role';
         $role = Role::findOrFail($id);
+        Gate::authorize('deleteAnyRoles', Role::class);
 
         
 
@@ -107,7 +110,7 @@ class RoleController extends Controller
     public function destroy($id)//
     {
         $role = Role::findOrFail($id);
-
+        Gate::authorize('deleteAnyRoles', Role::class);
         $role->delete();
         return redirect()->route('roles.index');
     }
